+++ NEWS.md
@@ -5,8 +5,9 @@
 ## New Features
 
 -   New `train_wordvec()` function (still in development).
+-   New `tokenize()` function.
 -   New `data_wordvec_reshape()` function.
--   New `tab_RND()` function.
+-   New `test_RND()` function (and `tab_WEAT()` is renamed to `test_WEAT()`).
 -   New `plot_wordvec_tSNE()` function: Visualize 2-D or 3-D word vectors with dimensionality reduced using the t-Distributed Stochastic Neighbor Embedding (t-SNE) method.
 
 # PsychWordVec 0.0.6 (Jul 2022)
@@ -16,7 +17,7 @@
 -   Enhanced all functions.
 -   New `data_wordvec_subset()` function.
 -   Added the `unique` argument for `tab_similarity()`.
--   Added support for using regular expression pattern in `tab_WEAT()`.
+-   Added support for using regular expression pattern in `test_WEAT()`.
 
 # PsychWordVec 0.0.4 (Apr 2022)